fix: firefox white screen with referer header
