+++ src/app/common/token-utils.ts
@@ -1,4 +1,5 @@
-import { abbreviateNumber, isUndefined } from '@app/common/utils';
+import { abbreviateNumber } from '@app/common/utils';
+import { isUndefined } from '@shared/utils';
 import { AssetWithMeta, FtMeta } from './asset-types';
 import { convertUnicodeToAscii } from './string-utils';
 import { isValidUrl } from './validation/validate-url';
