+++ src/app/common/validation/use-fee-schema.ts
@@ -5,7 +5,7 @@ import { STX_DECIMALS } from '@shared/constants';
 import { stxAmountSchema } from '@app/common/validation/currency-schema';
 import { formatInsufficientBalanceError, formatPrecisionError } from '@app/common/error-formatters';
 import { SendFormErrorMessages } from '@app/common/error-messages';
-import { isNumber } from '@app/common/utils';
+import { isNumber } from '@shared/utils';
 import { useCurrentAccountAvailableStxBalance } from '@app/store/accounts/account.hooks';
 import { stxToMicroStx } from '@app/common/stacks-utils';
