+++ src/app/common/validation/validate-memo.ts
@@ -1,6 +1,7 @@
 import * as yup from 'yup';
 import { MEMO_MAX_LENGTH_BYTES } from '@stacks/transactions';
-import { isString } from '@app/common/utils';
+
+import { isString } from '@shared/utils';
 
 const exceedsMaxLengthBytes = (string: string, maxLengthBytes: number): boolean =>
   string ? Buffer.from(string).length > maxLengthBytes : false;
