+++ src/app/features/ledger/ledger-jwt-signing.context.ts
@@ -1,6 +1,6 @@
 import { createContext } from 'react';
 
-import { noop } from '@app/common/utils';
+import { noop } from '@shared/utils';
 import { getAppVersion } from './ledger-utils';
 
 export interface LedgerJwtSigningProvider {
