+++ src/app/pages/send-tokens/components/send-max-button.tsx
@@ -4,7 +4,7 @@ import { Box, ButtonProps, color } from '@stacks/ui';
 
 import { AssetWithMeta } from '@app/common/asset-types';
 import { SendFormSelectors } from '@tests/page-objects/send-form.selectors';
-import { isUndefined } from '@app/common/utils';
+import { isUndefined } from '@shared/utils';
 
 function SendMaxButtonAction(props: ButtonProps) {
   return (
