+++ src/shared/utils.ts
@@ -1,0 +1,17 @@
+export function isNumber(value: unknown): value is number {
+  return typeof value === 'number';
+}
+
+export function isString(value: unknown): value is string {
+  return typeof value === 'string';
+}
+
+export function isUndefined(value: unknown): value is undefined {
+  return typeof value === 'undefined';
+}
+
+export function isEmpty(value: Object) {
+  return Object.keys(value).length === 0;
+}
+
+export function noop() {}
